Improving documentation and sample projects.
