+++ README.md
@@ -41,7 +41,7 @@
 1. We've prepackaged this project with a few starter templates, located [here](https://github.com/blitzstudios/sleeper-mini/tree/main/template/src).
 2. Select a project to start with by editing [app.json](https://github.com/blitzstudios/sleeper-mini/blob/6938294b47fb4f764f9e5e70e8b7d00749f38b4e/template/app.json#L6) and changing the selected sample ID. Restart your packager each time you change this value.
 3. All user changes should hook into the main index.js file of the sample you selected from [these folders](https://github.com/blitzstudios/sleeper-mini/tree/main/template/src). Feel free to delete or modify components as necessary, but refrain from modifying any library files or [index.tsx](https://github.com/blitzstudios/sleeper-mini/blob/main/template/index.tsx), as those handle some behind the scenes communication events.
-4. Feel free to edit the top level `package.json` and add new packages to your project. Note that these packages either must already exist within Sleeper (if they include native code), or must be entirely comprised of javascript. Contact us for an exhaustive list of supported native packages, or request new ones be added.
+4. Feel free to edit the top level `package.json` and add new packages to your project. Note that these packages either must already exist within Sleeper (if they include native code), or must be entirely comprised of javascript. For a list of packages included in Sleeper, check out [mini_packages.json](https://github.com/blitzstudios/sleeper-mini/blob/main/template/mini_packages.json).
 5. Some types and components have been provided for use through an included `@sleeperhq/mini-core` package. You can check out the repo [here](https://github.com/blitzstudios/sleeper-mini-core).
 6. Further documentation and API hooks are in development.
