+++ template/app.json
@@ -2,6 +2,6 @@
   "name": "template",
   "displayName": "template",
   "remoteIP": "192.168.86.190",
-  "samples": ["Mini", "Fetch", "Actions"],
+  "samples": ["Mini", "FetchSample", "ActionSample"],
   "selectedSample": 0
 }
