+++ template/mini_packages.json
@@ -1,0 +1,103 @@
+{
+  "availablePackageVersions": {
+    "@mattermost/react-native-paste-input": "0.4.0",
+    "@notifee/react-native": "7.0.1",
+    "@ptomasroos/react-native-multi-slider": "2.2.2",
+    "@react-native-camera-roll/camera-roll": "5.2.0",
+    "@react-native-community/blur": "3.6.0",
+    "@react-native-community/datetimepicker": "2.6.2",
+    "@react-native-community/hooks": "2.8.0",
+    "@react-native-community/netinfo": "9.3.7",
+    "@react-native-masked-view/masked-view": "0.2.8",
+    "@react-native-picker/picker": "1.16.8",
+    "@react-navigation/bottom-tabs": "6.5.4",
+    "@react-navigation/compat": "5.3.20",
+    "@react-navigation/native": "6.1.3",
+    "@react-navigation/stack": "6.3.12",
+    "@shopify/flash-list": "1.4.1",
+    "crypto-js": "3.3.0",
+    "decimal.js-light": "2.5.1",
+    "deep-diff": "1.0.2",
+    "easy-lottie-react-native": "2.0.2",
+    "lottie-react-native": "4.1.3",
+    "path": "0.12.7",
+    "react": "17.0.2",
+    "react-freeze": "1.0.3",
+    "react-native": "0.66.5",
+    "react-native-action-sheet": "2.2.0",
+    "react-native-animatable": "1.3.3",
+    "react-native-branch": "5.4.0",
+    "react-native-check-version": "1.0.20",
+    "react-native-color-matrix-image-filters": "5.2.10",
+    "react-native-compressor": "1.5.2",
+    "react-native-confirmation-code-field": "6.7.0",
+    "react-native-contacts": "5.2.8",
+    "react-native-create-thumbnail": "2.0.0-rc.2",
+    "react-native-date-picker": "4.2.2",
+    "react-native-detector": "0.2.1",
+    "react-native-device-info": "8.4.1",
+    "react-native-draggable-flatlist": "3.1.2",
+    "react-native-draggable-reanimated": "1.0.5",
+    "react-native-dropdownalert": "4.3.0",
+    "react-native-emoji-rain": "0.0.1",
+    "react-native-extra-dimensions-android": "1.2.5",
+    "react-native-fs": "2.18.0",
+    "react-native-gesture-handler": "2.8.0",
+    "react-native-google-places-autocomplete": "2.4.1",
+    "react-native-haptic-feedback": "1.6.0",
+    "react-native-image-crop-picker": "0.39.0",
+    "react-native-image-pan-zoom": "2.1.12",
+    "react-native-image-picker": "4.10.3",
+    "react-native-image-resizer": "1.4.5",
+    "react-native-iphone-x-helper": "1.3.1",
+    "react-native-json-tree": "1.3.0",
+    "react-native-keep-awake": "4.0.0",
+    "react-native-keyboard-aware-scroll-view": "0.9.5",
+    "react-native-keyboard-spacer": "0.4.1",
+    "react-native-linear-gradient": "2.5.6",
+    "react-native-modal": "13.0.0",
+    "react-native-modal-datetime-picker": "13.0.1",
+    "react-native-parsed-text": "0.0.21",
+    "react-native-permissions": "3.6.1",
+    "react-native-platform-touchable": "1.1.1",
+    "react-native-popup-dialog": "0.11.46",
+    "react-native-popup-menu": "0.16.1",
+    "react-native-progress": "4.1.2",
+    "react-native-qrcode-svg": "6.1.1",
+    "react-native-radar": "3.2.3",
+    "react-native-reanimated": "2.14.4",
+    "react-native-redash": "18.1.0",
+    "react-native-safe-area-context": "3.3.2",
+    "react-native-screens": "3.20.0",
+    "react-native-search-bar": "3.5.1",
+    "react-native-section-list-get-item-layout": "2.2.3",
+    "react-native-shake": "5.3.1",
+    "react-native-share": "7.3.9",
+    "react-native-shimmer-placeholder": "1.0.36",
+    "react-native-simple-toast": "1.1.3",
+    "react-native-sms": "1.11.0",
+    "react-native-snap-carousel": "3.9.1",
+    "react-native-sound": "0.11.1",
+    "react-native-sound-level": "1.1.5",
+    "react-native-spinkit": "1.5.1",
+    "react-native-svg": "13.7.0",
+    "react-native-swiper": "1.6.0-rc.3",
+    "react-native-tab-view": "3.5.0",
+    "react-native-tcp-socket": "6.0.6",
+    "react-native-triangle": "0.0.9",
+    "react-native-vector-icons": "9.0.0",
+    "react-native-video": "5.2.1",
+    "react-native-view-shot": "3.5.0",
+    "react-native-webview": "11.23.1",
+    "react-native-webp-format": "1.1.2",
+    "recyclerlistview": "4.2.0",
+    "rn-prompt": "1.0.4",
+    "@react-native-async-storage/async-storage": "1.15.15",
+    "@react-native-community/audio-toolkit": "2.0.3",
+    "react-native-chart-kit": "6.7.0",
+    "react-native-circular-progress": "1.3.6",
+    "react-native-fast-image": "8.5.11",
+    "react-native-interactable": "0.1.10",
+    "react-native-recyclerview-list": "0.3.4"
+  }
+}
